Ticket #415: fix implicit cast which makes Windows build fail.

+++ tests/demo/test_session.c
@@ -44,7 +44,7 @@ START_TEST(test_username_password)
     ck_assert(SOPC_Toolkit_Configured() == SOPC_STATUS_OK);
 
     ck_assert(StateMachine_StartSession_UsernamePassword(g_pSM, "UserName", "user", (const uint8_t*) "password",
-                                                         strlen("password")) == SOPC_STATUS_OK);
+                                                         (int32_t) strlen("password")) == SOPC_STATUS_OK);
     wait_for_machines(1, g_pSM);
 
     SOPC_Toolkit_Clear();
